+++ tests/test_txmonitor.py
@@ -11,6 +11,7 @@ from web3 import Web3, EthereumTesterProvider
 from web3._utils.transactions import fill_nonce
 
 from smart_contracts_for_testing.gas import estimate_gas_fees, GasPriceMethod, apply_gas
+from smart_contracts_for_testing.hotwallet import HotWallet
 from smart_contracts_for_testing.token import create_token
 from smart_contracts_for_testing.txmonitor import wait_transactions_to_complete, \
     broadcast_and_wait_transactions_to_complete
@@ -51,15 +52,18 @@ def hot_wallet_private_key(web3) -> HexBytes:
 
 
 @pytest.fixture()
-def hot_wallet(web3, hot_wallet_private_key) -> LocalAccount:
+def hot_wallet(web3, hot_wallet_private_key) -> HotWallet:
     """User account.
 
     Do some account allocation for tests.
     """
-    return Account.from_key(hot_wallet_private_key)
+    account = Account.from_key(hot_wallet_private_key)
+    hot_wallet = HotWallet(account)
+    hot_wallet.sync_nonce(web3)
+    return hot_wallet
 
 
-def test_wait_txs_parallel(web3: Web3, eth_tester, deployer: HexAddress, hot_wallet: LocalAccount):
+def test_wait_txs_parallel(web3: Web3, eth_tester, deployer: HexAddress, hot_wallet: HotWallet):
     """Wait multiple transactions to complete in parallel."""
 
     gas_fees = estimate_gas_fees(web3)
@@ -70,8 +74,6 @@ def test_wait_txs_parallel(web3: Web3, eth_tester, deployer: HexAddress, hot_wal
     web3.eth.send_transaction({"from": deployer, "to": hot_wallet.address, "value": 1 * 10**18})
     token.functions.transfer(hot_wallet.address, 50_000 * 10**18).transact({"from": deployer})
 
-    nonce = web3.eth.get_transaction_count(hot_wallet.address)
-
     # Create a raw transaction
     # Move 10 tokens from deployer to user1
     # https://web3py.readthedocs.io/en/stable/contracts.html?highlight=buildTransaction#web3.contract.ContractFunction.buildTransaction
@@ -79,21 +81,19 @@ def test_wait_txs_parallel(web3: Web3, eth_tester, deployer: HexAddress, hot_wal
         "from": hot_wallet.address,
         'chainId': web3.eth.chain_id,
         "gas": 150_000,  # 150k gas should be more than enough for ERC20.transfer(),
-        "nonce": nonce,
     })
 
     tx2 = token.functions.transfer(hot_wallet.address, 10 * 10**18).buildTransaction({
         "from": hot_wallet.address,
         'chainId': web3.eth.chain_id,
         "gas": 150_000,  # 150k gas should be more than enough for ERC20.transfer()
-        "nonce": nonce + 1,
     })
 
     apply_gas(tx1, gas_fees)
     apply_gas(tx2, gas_fees)
 
-    signed1 = hot_wallet.sign_transaction(tx1)
-    signed2 = hot_wallet.sign_transaction(tx2)
+    signed1 = hot_wallet.sign_transaction_with_new_nonce(tx1)
+    signed2 = hot_wallet.sign_transaction_with_new_nonce(tx2)
 
     tx_hash1 = web3.eth.send_raw_transaction(signed1.rawTransaction)
     tx_hash2 = web3.eth.send_raw_transaction(signed2.rawTransaction)
@@ -105,7 +105,7 @@ def test_wait_txs_parallel(web3: Web3, eth_tester, deployer: HexAddress, hot_wal
         assert receipt.status == 1  # tx success
 
 
-def test_broadcast_and_wait(web3: Web3, eth_tester, deployer: HexAddress, hot_wallet: LocalAccount):
+def test_broadcast_and_wait(web3: Web3, eth_tester, deployer: HexAddress, hot_wallet: HotWallet):
     """Broadcast and multiple transactions to complete in parallel."""
 
     gas_fees = estimate_gas_fees(web3)
@@ -116,8 +116,6 @@ def test_broadcast_and_wait(web3: Web3, eth_tester, deployer: HexAddress, hot_wa
     web3.eth.send_transaction({"from": deployer, "to": hot_wallet.address, "value": 1 * 10**18})
     token.functions.transfer(hot_wallet.address, 50_000 * 10**18).transact({"from": deployer})
 
-    nonce = web3.eth.get_transaction_count(hot_wallet.address)
-
     # Create a raw transaction
     # Move 10 tokens from deployer to user1
     # https://web3py.readthedocs.io/en/stable/contracts.html?highlight=buildTransaction#web3.contract.ContractFunction.buildTransaction
@@ -125,21 +123,19 @@ def test_broadcast_and_wait(web3: Web3, eth_tester, deployer: HexAddress, hot_wa
         "from": hot_wallet.address,
         'chainId': web3.eth.chain_id,
         "gas": 150_000,  # 150k gas should be more than enough for ERC20.transfer(),
-        "nonce": nonce,
     })
 
     tx2 = token.functions.transfer(hot_wallet.address, 10 * 10**18).buildTransaction({
         "from": hot_wallet.address,
         'chainId': web3.eth.chain_id,
         "gas": 150_000,  # 150k gas should be more than enough for ERC20.transfer()
-        "nonce": nonce + 1,
     })
 
     apply_gas(tx1, gas_fees)
     apply_gas(tx2, gas_fees)
 
-    signed1 = hot_wallet.sign_transaction(tx1)
-    signed2 = hot_wallet.sign_transaction(tx2)
+    signed1 = hot_wallet.sign_transaction_with_new_nonce(tx1)
+    signed2 = hot_wallet.sign_transaction_with_new_nonce(tx2)
 
     complete = broadcast_and_wait_transactions_to_complete(web3, [signed1, signed2])